Add comment cleanup script and update various code comments
Added CleanComments.ps1 PowerShell script to remove specific comment prefixes from source files. Updated project files to include the script. Refactored and cleaned up comments across multiple source and shader files, removing 'CRITICAL', 'FIX', and similar prefixes for clarity. Updated README and scene snapshot for new defaults and minor text changes. Adjusted ImGui window positions and made minor code and documentation improvements throughout the codebase.

+++ src/GuiNode.cpp
@@ -441,7 +441,7 @@ void GuiNode::RenderText(const GuiElement& e) const {
         e.color.b / 255.0f,
         e.color.a / 255.0f
     );
-    drawQuad(tex, e.x, e.y, converted->w, converted->h, tint);
+    drawQuad(tex, e.x, e.y, static_cast<float>(converted->w), static_cast<float>(converted->h), tint);
 
     glDeleteTextures(1, &tex);
     SDL_FreeSurface(converted);
@@ -495,7 +495,7 @@ void GuiNode::RenderProgressBar(const GuiElement& e) const {
                 float labelY = e.y + (e.height - converted->h) * 0.5f;
                 
                 glm::vec4 tint(1.0f, 1.0f, 1.0f, 1.0f);
-                drawQuad(tex, labelX, labelY, converted->w, converted->h, tint);
+                drawQuad(tex, labelX, labelY, static_cast<float>(converted->w), static_cast<float>(converted->h), tint);
                 
                 glDeleteTextures(1, &tex);
                 SDL_FreeSurface(converted);
